+++ src/services/handler.service.js
@@ -25,42 +25,26 @@ class Handler {
   }
 
   async handle(target, name, active) {
-    let result = {
-      error: true,
-      message: 'Homebridge not initialized.',
-    };
-
     if (this.initialized) {
       const accessory = this.accessories.find((accessory) => accessory.displayName == name);
 
       if (accessory) {
         switch (target) {
           case 'motion':
-            result = await this.motionHandler(accessory, active);
+            await this.motionHandler(accessory, active);
             break;
           case 'doorbell':
-            result = await this.doorbellHandler(accessory, active);
+            await this.doorbellHandler(accessory, active);
             break;
           default:
-            result = {
-              error: true,
-              message: `First directory level must be "motion" or "doorbell", got "${target}".`,
-            };
+            this.log.error(`Unknown target specified for motion handler (${target})!`, name, 'Homebridge');
         }
       } else {
-        result = {
-          error: true,
-          message: `Camera "${name}" not found.`,
-        };
+        this.log.error(`Camera "${name}" not found.`);
       }
+    } else {
+      this.log.error('Homebridge not initialized.', name, 'Homebridge');
     }
-
-    if (result.error) {
-      this.log.error(`Handling event: ${result.message}`, name, 'Homebridge');
-      return;
-    }
-
-    this.log.debug(`Handling event: ${result.message}`, name);
   }
 
   async motionHandler(accessory, active, manual) {
@@ -74,16 +58,13 @@ class Handler {
 
       if (timeout) {
         if (active) {
-          this.log.info('Motion ON (skip motion event, timeout active)', accessory.displayName);
+          this.log.info('Motion ON - Skip motion event, timeout active!', accessory.displayName);
 
           if (motionTrigger && manual) {
             setTimeout(() => motionTrigger.updateCharacteristic(this.hap.Characteristic.On, false), 500);
           }
 
-          return {
-            error: false,
-            message: 'Skip motion event, timeout active!',
-          };
+          return;
         } else {
           clearTimeout(timeout);
           this.motionTimers.delete(accessory.UUID);
@@ -97,7 +78,7 @@ class Handler {
 
         if (active && atHome && !cameraExcluded) {
           this.log.info(
-            `Motion ON (skip motion event, at home is active and ${accessory.displayName} is not excluded)`,
+            `Motion ON - Skip motion trigger. At Home is active and ${accessory.displayName} is not excluded!`,
             accessory.displayName
           );
 
@@ -105,17 +86,14 @@ class Handler {
             setTimeout(() => motionTrigger.updateCharacteristic(this.hap.Characteristic.On, false), 500);
           }
 
-          return {
-            error: false,
-            message: `Skip motion trigger. At Home is active and ${accessory.displayName} is not excluded!`,
-          };
+          return;
         }
+      }
 
-        this.log.info(`Motion ${active ? 'ON' : 'OFF'}`, accessory.displayName);
+      this.log.info(`Motion ${active ? 'ON' : 'OFF'}`, accessory.displayName);
 
-        if (!accessory.context.config.hsv) {
-          this.cameraUi.eventController.triggerEvent('motion', accessory.displayName, active);
-        }
+      if (!accessory.context.config.hsv && manual) {
+        this.cameraUi.eventController.triggerEvent('motion', accessory.displayName, active);
       }
 
       motionSensor.updateCharacteristic(this.hap.Characteristic.MotionDetected, active ? true : false);
@@ -130,7 +108,7 @@ class Handler {
         }
 
         const timer = setTimeout(() => {
-          this.log.info('Motion handler timeout.', accessory.displayName);
+          this.log.info('Motion OFF - Motion handler timeout.', accessory.displayName);
 
           this.motionTimers.delete(accessory.UUID);
           motionSensor.updateCharacteristic(this.hap.Characteristic.MotionDetected, false);
@@ -142,20 +120,12 @@ class Handler {
 
         this.motionTimers.set(accessory.UUID, timer);
       }
-
-      return {
-        error: false,
-        message: `Motion switched ${active ? 'on' : 'off'}`,
-      };
     } else {
       if (motionTrigger) {
         setTimeout(() => motionTrigger.updateCharacteristic(this.hap.Characteristic.On, false), 500);
       }
 
-      return {
-        error: true,
-        message: 'Motion is not enabled for this camera.',
-      };
+      this.log.debug('Motion is not enabled for this camera.');
     }
   }
 
@@ -169,16 +139,13 @@ class Handler {
 
       if (timeout) {
         if (active) {
-          this.log.info('Doorbell ON (skip doorbell event, doorbell timeout active)', accessory.displayName);
+          this.log.info('Doorbell ON - Skip doorbell event, timeout active!', accessory.displayName);
 
           if (doorbellTrigger && manual) {
             setTimeout(() => doorbellTrigger.updateCharacteristic(this.hap.Characteristic.On, false), 500);
           }
 
-          return {
-            error: false,
-            message: 'Skip doorbell event, timeout active!',
-          };
+          return;
         } else {
           clearTimeout(timeout);
           this.doorbellTimers.delete(accessory.UUID);
@@ -192,7 +159,7 @@ class Handler {
 
         if (active && atHome && !cameraExcluded) {
           this.log.info(
-            `Doorbell ON (skip doorbell event, at home is active and ${accessory.displayName} is not excluded)`,
+            `Doorbell ON  - Skip doorbell trigger. At Home is active and ${accessory.displayName} is not excluded!`,
             accessory.displayName
           );
 
@@ -200,18 +167,15 @@ class Handler {
             setTimeout(() => doorbellTrigger.updateCharacteristic(this.hap.Characteristic.On, false), 500);
           }
 
-          return {
-            error: false,
-            message: `Skip doorbell trigger. At Home is active and ${accessory.displayName} is not excluded!`,
-          };
-        }
-
-        if (!accessory.context.config.hsv) {
-          this.cameraUi.eventController.triggerEvent('doorbell', accessory.displayName, active);
+          return;
         }
       }
 
       this.log.info(`Doorbell ${active ? 'ON' : 'OFF'}`, accessory.displayName);
+
+      if (!accessory.context.config.hsv && manual) {
+        this.cameraUi.eventController.triggerEvent('doorbell', accessory.displayName, active);
+      }
 
       doorbellTrigger?.updateCharacteristic(this.hap.Characteristic.On, active ? true : false);
 
@@ -226,7 +190,7 @@ class Handler {
         );
 
         const timer = setTimeout(() => {
-          this.log.debug('Doorbell handler timeout.', accessory.displayName);
+          this.log.info('Doorbell OFF - Doorbell handler timeout.', accessory.displayName);
 
           this.doorbellTimers.delete(accessory.UUID);
 
@@ -237,20 +201,12 @@ class Handler {
 
         this.doorbellTimers.set(accessory.UUID, timer);
       }
-
-      return {
-        error: false,
-        message: `Doorbell switched ${active ? 'on' : 'off'}`,
-      };
     } else {
       if (doorbellTrigger) {
         setTimeout(() => doorbellTrigger.updateCharacteristic(this.hap.Characteristic.On, false), 500);
       }
 
-      return {
-        error: true,
-        message: 'Doorbell is not enabled for this camera.',
-      };
+      this.log.debug('Doorbell is not enabled for this camera.');
     }
   }
 }